requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
Pillow>=8.0.0
scipy>=1.6.0
